requirement.txt
simplejson
fvcore
pytorchvideo
matplotlib
tensorboardX
opencv-python
pytorch-gradcam
scikit-learn
prettytable